+++ packages/featureserver/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* check isNaN instead of falsy for extent validity
+
 ## [2.4.1] -08-22-2017
 ### Fixed
 * Don't bomb out /FeatureServer calls when metadata with no extent is passed in